requirements.txt
pandas
matplotlib
yfinance
wordcloud
requests
python-dotenv
praw
transformers
torch
huggingface_hub[hf_xet]
nltk
scikit-learn